Subcurve are now  updated with the right information

+++ src/Core/Arrangement_2D/Arrangement_2D.hpp
@@ -17,7 +17,7 @@
 #include "PolygonalCurve2.hpp"
 #include "Core/Geometry/PolygonalCurve/PolygonalCurve2D.hpp"
 
-typedef unsigned long long Natural;
+typedef unsigned int Natural;
 typedef double Real;
 typedef RRM::PolygonalCurve2 PolyCurve2D;
